risk/requirements.txt
amqp==2.1.4
aniso8601==1.2.1
BestSignAPI==1.1.0
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
click==6.7
contextlib2==0.5.5
Flask==0.12.2
Flask-Login==0.4.0
Flask-RESTful==0.3.6
gevent==1.2.2
geventhttpclient==1.3.1
greenlet==0.4.12
grequests==0.3.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
mongoengine==0.13.0
mysql-connector-python-rf==2.2.2
olefile==0.44
Pillow==4.2.1
pycrypto==2.6.1
pymongo==3.4.0
python-dateutil==2.6.1
pytz==2017.2
qiniu==7.1.5
qrcode==5.3
raven==6.1.0
redis==2.10.5
requests==2.18.1
six==1.10.0
SQLAlchemy==1.1.11
urllib3==1.21.1
uWSGI==2.0.15
vine==1.1.3
Werkzeug==0.12.2
xlrd==1.1.0
XlsxWriter==0.9.8
pymysql==0.7.11
blinker==1.4
jieba==0.39
